Refactor and minor fixes across UI and editor components
Removed unused variables and improved code clarity in several components. Updated theme naming in TSettings, fixed Q_PROPERTY usage in TFlatButton, and corrected regular expression match options in TLexer. Applied Q_UNUSED to unused parameters and improved bounds checking in AutoCompleteUI.

+++ source/texteditor/autocomplete/AutoCompleteUI.cpp
@@ -21,7 +21,7 @@ int CompletionModel::rowCount(const QModelIndex &) const {
 }
 
 QVariant CompletionModel::data(const QModelIndex &index, int role) const {
-    if (!index.isValid() || index.row() >= m_data.size()) return QVariant();
+    if (!index.isValid() || static_cast<size_t>(index.row()) >= m_data.size()) return QVariant();
     const auto &item = m_data[index.row()];
 
     if (role == Qt::DisplayRole) return item.label;
